pyproject.toml
[tool.poetry]
name = "aline"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.9,<3.12"
Flask = "^2.2.2"
python-dotenv = "^0.21.0"
gunicorn = "^20.1.0"
readability-lxml = "^0.8.1"
lxml = "^4.9.2"
requests = "^2.28.2"
openai = "^0.27.0"
langchain = "^0.0.139"
markdownify = "^0.11.6"
mistletoe = "^1.0.1"
tiktoken = "^0.3.0"
pinecone-client = "^2.2.1"
chromadb = "^0.3.11"
modal-client = "^0.48.1862"
anthropic = "^0.2.7"
boto3 = "^1.26.114"
sentence-transformers = "^2.2.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
ipykernel = "^6.21.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
